requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kubernetes==32.0.0
lz4==4.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
phidata==2.7.10
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
posthog==3.11.0
proto-plus==1.26.0
protobuf
protoc-gen-openapiv2==0.0.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pylance==0.22.0
pyparsing==3.2.1
pypdf==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tantivy==0.22.0
tavily-python==0.5.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
zipp==3.21.0
